requirements.txt
# File: worker/requirements.txt
flask
pdfplumber
pytesseract
Pillow
requests
python-dotenv
opencv-python-headless
